refactor: ManyToMany 를 위한 중간자 테이블이 저장되는 패키지 이름을 변경합니다
